+++ NEWS.md
@@ -1,5 +1,6 @@
 # idaifieldR 0.2.2 _2022_
 * config specific names now removed everywhere 
+* make `find_layer()` faster by replacing `which()` with `match()` and avoiding `rbind()`
 * more tests
 
 # idaifieldR 0.2.1 _2022-10-23_